+++ src/QueryType/Server/Collections/Result/ClusterStatusResult.php
@@ -15,6 +15,7 @@ use Solarium\QueryType\Server\Query\AbstractResult;
 /**
  * ClusterStatusResult.
  */
+#[\AllowDynamicProperties]
 class ClusterStatusResult extends AbstractResult
 {
     /**