+++ src/QueryType/Server/Collections/Result/CreateResult.php
@@ -14,6 +14,7 @@ use Solarium\QueryType\Server\Query\AbstractResult;
 /**
  * CreateResult.
  */
+#[\AllowDynamicProperties]
 class CreateResult extends AbstractResult
 {
     /**