+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the Solarium library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- PHP 8.2 deprecations for Solarium\QueryType\Server\Collections results
+
+
 ## [6.3.4]
 ### Added
 - PHP 8.3 support